move defaults e strategies para raiz e ajusta imports gerais
